sending some stuff overr

+++ ardunio/serial_test.ino
@@ -1,6 +1,7 @@
 #include <Arduino.h>
 #include <HardwareSerial.h>
-// #include "../shared/CommunicationsConstants.h"
+// #include "../shared/CommunicationsConstants.h"]
+#include <cstdint>
 
 void setup()
 {
@@ -11,7 +12,7 @@ void loop()
 {
     if (Serial.available() > 0)
     {
-        int bt = Serial.read();
+        uint8_t bt = Serial.read();
 
         Serial.write(bt == 0 ? 8 : bt == 1 ? 34 : 10);
     }
